+++ src/chefe/shared/__init__.py
@@ -1,1 +1,2 @@
 from .environment import GlobalEnv as GlobalEnv
+from .stage import SecondStage as SecondStage